:sparkles:Feat: week05 기능 추가

+++ src/main/java/com/likelion/springpractice/domain/post/week04/entity/Post.java
@@ -31,6 +31,6 @@ public class Post extends BaseTimeEntity {
   private String content; // 내용
 
   @Column
-  private Long views; // 조회수
+  private int views; // 조회수
 
 }
